+++ WalletWasabi.Gui/Program.cs
@@ -1,23 +1,4 @@
-using Avalonia;
-using Avalonia.Dialogs;
-using Avalonia.ReactiveUI;
-using Avalonia.Threading;
-using AvalonStudio.Extensibility;
-using AvalonStudio.Shell;
-using AvalonStudio.Shell.Extensibility.Platforms;
-using Splat;
-using System;
-using System.IO;
 using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
-using System.Threading.Tasks;
-using WalletWasabi.Gui.CommandLine;
-using WalletWasabi.Gui.CrashReport;
-using WalletWasabi.Gui.ViewModels;
-using WalletWasabi.Helpers;
-using WalletWasabi.Logging;
-using WalletWasabi.Services.Terminate;
-using WalletWasabi.Wallets;
 
 // This is temporary and to facilitate the migration to new UI.
 [assembly: InternalsVisibleTo("WalletWasabi.Fluent")]
@@ -31,9 +12,8 @@ namespace WalletWasabi.Gui
 		/// see https://github.com/AvaloniaUI/Avalonia/wiki/Unresolved-platform-support-issues
 		private static void Main(string[] args)
 		{
-			GuiProgram program = new GuiProgram();
+			GuiProgram program = new ();
 			program.Run(args);
 		}
-
 	}
 }
